refactor(ipc): split worktree.ts into focused sub-domain handlers (#5240)
- Split 882-LOC electron/ipc/handlers/worktree.ts into focused modules
  under electron/ipc/handlers/worktree/: lifecycle, branches, pull-requests,
  task, plus shared constants.ts and a composer index.ts
- Extract 4 git-read handlers (GIT_COMPARE_WORKTREES, GIT_GET_FILE_DIFF,
  GIT_GET_PROJECT_PULSE, GIT_LIST_COMMITS) into new sibling
  electron/ipc/handlers/git-read.ts alongside git-write.ts
- Replace worktree.ts with a deprecated re-export shim (mirrors terminal.ts
  precedent) so existing test imports continue to resolve
- Register registerGitReadHandlers in handlers.ts and the registry test mock
- No IPC channel renames; no behavioral changes
