+++ electron/ipc/handlers.ts
@@ -35,6 +35,7 @@ import { registerCliHandlers } from "./handlers/cli.js";
 import { registerHelpHandlers } from "./handlers/help.js";
 import { registerClipboardHandlers } from "./handlers/clipboard.js";
 import { registerGitWriteHandlers } from "./handlers/git-write.js";
+import { registerGitReadHandlers } from "./handlers/git-read.js";
 import { registerTelemetryHandlers } from "./handlers/telemetry.js";
 import { registerPrivacyHandlers } from "./handlers/privacy.js";
 import { registerSentryHandlers } from "./handlers/sentry.js";
@@ -130,6 +131,7 @@ export function registerIpcHandlers(deps: HandlerDependencies): () => void {
     register(() => registerHelpHandlers());
     register(() => registerClipboardHandlers());
     register(() => registerGitWriteHandlers(deps));
+    register(() => registerGitReadHandlers(deps));
     register(() => registerTelemetryHandlers());
     register(() => registerPrivacyHandlers());
     register(() => registerSentryHandlers());
