+++ electron/ipc/handlers/worktree.ts
@@ -1,882 +1,8 @@
-import { ipcMain } from "electron";
-import { getWindowForWebContents } from "../../window/webContentsRegistry.js";
-import path from "path";
-import { CHANNELS } from "../channels.js";
-import { store } from "../../store.js";
-import type { HandlerDependencies } from "../types.js";
-import type { WorktreeSetActivePayload, WorktreeDeletePayload } from "../../types/index.js";
-import type { PulseRangeDays, ProjectPulse } from "../../../shared/types/pulse.js";
-import type {
-  CreateForTaskPayload,
-  CleanupTaskOptions,
-  AttachIssuePayload,
-  DetachIssuePayload,
-  IssueAssociation,
-} from "../../../shared/types/ipc/worktree.js";
-import type { WorktreeState } from "../../../shared/types/worktree.js";
-import { generateWorktreePath, validatePathPattern } from "../../../shared/utils/pathPattern.js";
-import { projectStore } from "../../services/ProjectStore.js";
-import { logDebug, logError } from "../../utils/logger.js";
-import { fileSearchService } from "../../services/FileSearchService.js";
-import { soundService } from "../../services/SoundService.js";
-import { checkRateLimit, waitForRateLimitSlot } from "../utils.js";
-
-const WORKTREE_RATE_LIMIT_KEY = "worktreeCreate";
-// Strict-interval leaky bucket: one worktree creation released every 1s.
-// Prior sliding-window approach (2 per 2s) released both slots simultaneously
-// when the window rolled, producing a feast/famine burst pattern. See #5098.
-// Interval reduced from 6s to 1s in #5161 — 1 per second is safe for git lock
-// contention on any repo size and brings 30-worktree batches down from ~180s
-// to ~30s without regressing the burst-pattern fix.
-const WORKTREE_RATE_LIMIT_INTERVAL_MS = 1_000;
-import { resolveWorktreePattern } from "../../utils/worktreePattern.js";
-import { taskWorktreeService } from "../../services/TaskWorktreeService.js";
-
-export function registerWorktreeHandlers(deps: HandlerDependencies): () => void {
-  const handlers: Array<() => void> = [];
-
-  const handleWorktreeGetAll = async (event: Electron.IpcMainInvokeEvent) => {
-    if (!deps.worktreeService) {
-      return [];
-    }
-    const senderWindow = getWindowForWebContents(event.sender);
-    return await deps.worktreeService.getAllStatesAsync(senderWindow?.id);
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_GET_ALL, handleWorktreeGetAll);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_GET_ALL));
-
-  const handleWorktreeRefresh = async () => {
-    if (!deps.worktreeService) {
-      return;
-    }
-    await deps.worktreeService.refresh();
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_REFRESH, handleWorktreeRefresh);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_REFRESH));
-
-  const handleWorktreePRRefresh = async () => {
-    if (!deps.worktreeService) {
-      return;
-    }
-    await deps.worktreeService.refreshPullRequests();
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_PR_REFRESH, handleWorktreePRRefresh);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_PR_REFRESH));
-
-  const handleWorktreePRStatus = async () => {
-    if (!deps.worktreeService) {
-      return null;
-    }
-    return await deps.worktreeService.getPRStatus();
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_PR_STATUS, handleWorktreePRStatus);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_PR_STATUS));
-
-  const handleWorktreeSetActive = async (
-    event: Electron.IpcMainInvokeEvent,
-    payload: WorktreeSetActivePayload
-  ) => {
-    if (!deps.worktreeService) {
-      return;
-    }
-    const senderWindow = getWindowForWebContents(event.sender);
-    const windowId = senderWindow?.id;
-    await deps.worktreeService.setActiveWorktree(payload.worktreeId, windowId, { silent: true });
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_SET_ACTIVE, handleWorktreeSetActive);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_SET_ACTIVE));
-
-  const handleWorktreeCreate = async (
-    _event: Electron.IpcMainInvokeEvent,
-    payload: {
-      rootPath: string;
-      options: { baseBranch: string; newBranch: string; path: string; fromRemote?: boolean };
-    }
-  ): Promise<string> => {
-    await waitForRateLimitSlot(WORKTREE_RATE_LIMIT_KEY, WORKTREE_RATE_LIMIT_INTERVAL_MS);
-    if (!deps.worktreeService) {
-      throw new Error("Workspace client not initialized");
-    }
-    const worktreeId = await deps.worktreeService.createWorktree(payload.rootPath, payload.options);
-    try {
-      fileSearchService.invalidate(payload.options.path);
-    } catch (error) {
-      console.warn("[worktree.create] Failed to invalidate file search cache:", error);
-    }
-    if (store.get("notificationSettings").uiFeedbackSoundEnabled) {
-      soundService.play("worktree-create");
-    }
-    return worktreeId;
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_CREATE, handleWorktreeCreate);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_CREATE));
-
-  const handleWorktreeListBranches = async (
-    _event: Electron.IpcMainInvokeEvent,
-    payload: { rootPath: string }
-  ) => {
-    if (!deps.worktreeService) {
-      throw new Error("Workspace client not initialized");
-    }
-    return await deps.worktreeService.listBranches(payload.rootPath);
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_LIST_BRANCHES, handleWorktreeListBranches);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_LIST_BRANCHES));
-
-  const handleWorktreeFetchPRBranch = async (
-    _event: Electron.IpcMainInvokeEvent,
-    payload: { rootPath: string; prNumber: number; headRefName: string }
-  ) => {
-    if (!deps.worktreeService) {
-      throw new Error("Workspace client not initialized");
-    }
-    if (!payload.rootPath || typeof payload.rootPath !== "string") {
-      throw new Error("rootPath is required");
-    }
-    if (!payload.prNumber || typeof payload.prNumber !== "number" || payload.prNumber <= 0) {
-      throw new Error("prNumber must be a positive number");
-    }
-    if (!payload.headRefName || typeof payload.headRefName !== "string") {
-      throw new Error("headRefName is required");
-    }
-    return await deps.worktreeService.fetchPRBranch(
-      payload.rootPath,
-      payload.prNumber,
-      payload.headRefName
-    );
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_FETCH_PR_BRANCH, handleWorktreeFetchPRBranch);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_FETCH_PR_BRANCH));
-
-  const handleWorktreeGetRecentBranches = async (
-    _event: Electron.IpcMainInvokeEvent,
-    payload: { rootPath: string }
-  ) => {
-    if (!deps.worktreeService) {
-      throw new Error("Workspace client not initialized");
-    }
-    return await deps.worktreeService.getRecentBranches(payload.rootPath);
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_GET_RECENT_BRANCHES, handleWorktreeGetRecentBranches);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_GET_RECENT_BRANCHES));
-
-  const handleWorktreeGetDefaultPath = async (
-    _event: Electron.IpcMainInvokeEvent,
-    payload: { rootPath: string; branchName: string }
-  ): Promise<string> => {
-    if (!payload || typeof payload !== "object") {
-      throw new Error("Invalid payload for worktree:get-default-path");
-    }
-
-    const { rootPath, branchName } = payload;
-
-    if (typeof rootPath !== "string" || !rootPath.trim()) {
-      throw new Error("Invalid rootPath: must be a non-empty string");
-    }
-
-    if (typeof branchName !== "string" || !branchName.trim()) {
-      throw new Error("Invalid branchName: must be a non-empty string");
-    }
-
-    const pattern = await resolveWorktreePattern(rootPath);
-
-    const validation = validatePathPattern(pattern);
-    if (!validation.valid) {
-      throw new Error(`Invalid stored pattern: ${validation.error}`);
-    }
-
-    // Generate the initial path
-    const initialPath = generateWorktreePath(rootPath, branchName, pattern);
-
-    // Auto-resolve path conflicts by finding an available path
-    const gitService = taskWorktreeService.getGitService(rootPath);
-    return gitService.findAvailablePath(initialPath);
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_GET_DEFAULT_PATH, handleWorktreeGetDefaultPath);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_GET_DEFAULT_PATH));
-
-  const handleWorktreeGetAvailableBranch = async (
-    _event: Electron.IpcMainInvokeEvent,
-    payload: { rootPath: string; branchName: string }
-  ): Promise<string> => {
-    if (!payload || typeof payload !== "object") {
-      throw new Error("Invalid payload for worktree:get-available-branch");
-    }
-
-    const { rootPath, branchName } = payload;
-
-    if (typeof rootPath !== "string" || !rootPath.trim()) {
-      throw new Error("Invalid rootPath: must be a non-empty string");
-    }
-
-    if (typeof branchName !== "string" || !branchName.trim()) {
-      throw new Error("Invalid branchName: must be a non-empty string");
-    }
-
-    const gitService = taskWorktreeService.getGitService(rootPath);
-    return gitService.findAvailableBranchName(branchName);
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_GET_AVAILABLE_BRANCH, handleWorktreeGetAvailableBranch);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_GET_AVAILABLE_BRANCH));
-
-  const handleWorktreeDelete = async (
-    event: Electron.IpcMainInvokeEvent,
-    payload: WorktreeDeletePayload
-  ) => {
-    checkRateLimit(CHANNELS.WORKTREE_DELETE, 10, 10_000);
-    if (!deps.worktreeService) {
-      throw new Error("Workspace client not initialized");
-    }
-    if (!payload || typeof payload !== "object") {
-      throw new Error("Invalid payload");
-    }
-    if (typeof payload.worktreeId !== "string" || !payload.worktreeId) {
-      throw new Error("Invalid worktree ID");
-    }
-    if (payload.force !== undefined && typeof payload.force !== "boolean") {
-      throw new Error("Invalid force parameter");
-    }
-    if (payload.deleteBranch !== undefined && typeof payload.deleteBranch !== "boolean") {
-      throw new Error("Invalid deleteBranch parameter");
-    }
-    const senderWindow = getWindowForWebContents(event.sender);
-    const states = await deps.worktreeService.getAllStatesAsync(senderWindow?.id);
-    const worktree = states.find((wt) => wt.id === payload.worktreeId);
-    await deps.worktreeService.deleteWorktree(
-      payload.worktreeId,
-      payload.force,
-      payload.deleteBranch
-    );
-    if (worktree) {
-      try {
-        fileSearchService.invalidate(worktree.path);
-      } catch (error) {
-        console.warn("[worktree.delete] Failed to invalidate file search cache:", error);
-      }
-    }
-    if (store.get("notificationSettings").uiFeedbackSoundEnabled) {
-      soundService.play("worktree-delete");
-    }
-    // Clean up persisted issue association
-    const issueMap = store.get("worktreeIssueMap", {});
-    if (issueMap[payload.worktreeId]) {
-      const { [payload.worktreeId]: _, ...rest } = issueMap;
-      store.set("worktreeIssueMap", rest);
-    }
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_DELETE, handleWorktreeDelete);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_DELETE));
-
-  const handleGitCompareWorktrees = async (
-    _event: Electron.IpcMainInvokeEvent,
-    payload: {
-      cwd: string;
-      branch1: string;
-      branch2: string;
-      filePath?: string;
-      useMergeBase?: boolean;
-    }
-  ) => {
-    checkRateLimit(CHANNELS.GIT_COMPARE_WORKTREES, 20, 10_000);
-    if (!payload || typeof payload !== "object") {
-      throw new Error("Invalid payload");
-    }
-
-    const { cwd, branch1, branch2, filePath, useMergeBase } = payload;
-
-    if (typeof cwd !== "string" || !cwd) {
-      throw new Error("Invalid working directory");
-    }
-    if (typeof branch1 !== "string" || !branch1) {
-      throw new Error("Invalid branch1");
-    }
-    if (typeof branch2 !== "string" || !branch2) {
-      throw new Error("Invalid branch2");
-    }
-    if (filePath !== undefined && (typeof filePath !== "string" || !filePath)) {
-      throw new Error("Invalid filePath");
-    }
-    if (useMergeBase !== undefined && typeof useMergeBase !== "boolean") {
-      throw new Error("Invalid useMergeBase");
-    }
-
-    const gitService = taskWorktreeService.getGitService(cwd);
-    return gitService.compareWorktrees(branch1, branch2, filePath, useMergeBase);
-  };
-  ipcMain.handle(CHANNELS.GIT_COMPARE_WORKTREES, handleGitCompareWorktrees);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.GIT_COMPARE_WORKTREES));
-
-  const handleGitGetFileDiff = async (
-    _event: Electron.IpcMainInvokeEvent,
-    payload: { cwd: string; filePath: string; status: string }
-  ): Promise<string> => {
-    checkRateLimit(CHANNELS.GIT_GET_FILE_DIFF, 10, 10_000);
-    if (!payload || typeof payload !== "object") {
-      throw new Error("Invalid payload");
-    }
-
-    const { cwd, filePath, status } = payload;
-
-    if (typeof cwd !== "string" || !cwd) {
-      throw new Error("Invalid working directory");
-    }
-    if (typeof filePath !== "string" || !filePath) {
-      throw new Error("Invalid file path");
-    }
-    if (typeof status !== "string" || !status) {
-      throw new Error("Invalid file status");
-    }
-
-    if (!deps.worktreeService) {
-      throw new Error("WorkspaceClient not initialized");
-    }
-
-    try {
-      return await deps.worktreeService.getFileDiff(cwd, filePath, status);
-    } catch (error) {
-      const errorMessage = error instanceof Error ? error.message : String(error);
-      console.error("[Git] Failed to get file diff via WorkspaceClient:", errorMessage);
-      throw new Error(`Failed to get file diff: ${errorMessage}`, { cause: error });
-    }
-  };
-  ipcMain.handle(CHANNELS.GIT_GET_FILE_DIFF, handleGitGetFileDiff);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.GIT_GET_FILE_DIFF));
-
-  const handleGitGetProjectPulse = async (
-    event: Electron.IpcMainInvokeEvent,
-    payload: {
-      worktreeId: string;
-      rangeDays: PulseRangeDays;
-      includeDelta?: boolean;
-      includeRecentCommits?: boolean;
-      forceRefresh?: boolean;
-    }
-  ): Promise<ProjectPulse> => {
-    checkRateLimit(CHANNELS.GIT_GET_PROJECT_PULSE, 10, 10_000);
-    if (!payload || typeof payload !== "object") {
-      throw new Error("Invalid payload");
-    }
-
-    const { worktreeId, rangeDays, includeDelta, includeRecentCommits, forceRefresh } = payload;
-
-    if (typeof worktreeId !== "string" || !worktreeId) {
-      throw new Error("Invalid worktree ID");
-    }
-
-    if (![60, 120, 180].includes(rangeDays)) {
-      throw new Error("Invalid rangeDays: must be 60, 120, or 180");
-    }
-
-    if (includeDelta !== undefined && typeof includeDelta !== "boolean") {
-      throw new Error("Invalid includeDelta: must be a boolean");
-    }
-
-    if (includeRecentCommits !== undefined && typeof includeRecentCommits !== "boolean") {
-      throw new Error("Invalid includeRecentCommits: must be a boolean");
-    }
-
-    if (forceRefresh !== undefined && typeof forceRefresh !== "boolean") {
-      throw new Error("Invalid forceRefresh: must be a boolean");
-    }
-
-    if (!deps.worktreeService) {
-      throw new Error("WorkspaceClient not initialized");
-    }
-
-    const monitor = await deps.worktreeService.getMonitorAsync(worktreeId);
-    if (!monitor) {
-      throw new Error(`Worktree not found: ${worktreeId}`);
-    }
-
-    const senderWindowPulse = getWindowForWebContents(event.sender);
-    const states = await deps.worktreeService.getAllStatesAsync(senderWindowPulse?.id);
-    const mainWorktree = states.find((wt) => wt.isMainWorktree);
-    const mainBranch = mainWorktree?.branch ?? "main";
-
-    return deps.worktreeService.getProjectPulse(monitor.path, worktreeId, mainBranch, rangeDays, {
-      includeDelta,
-      includeRecentCommits,
-      forceRefresh,
-    });
-  };
-  ipcMain.handle(CHANNELS.GIT_GET_PROJECT_PULSE, handleGitGetProjectPulse);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.GIT_GET_PROJECT_PULSE));
-
-  const handleGitListCommits = async (
-    _event: Electron.IpcMainInvokeEvent,
-    payload: {
-      cwd: string;
-      search?: string;
-      branch?: string;
-      skip?: number;
-      limit?: number;
-    }
-  ) => {
-    checkRateLimit(CHANNELS.GIT_LIST_COMMITS, 10, 10_000);
-    if (!payload || typeof payload !== "object") {
-      throw new Error("Invalid payload");
-    }
-
-    const { cwd, search, branch, skip, limit } = payload;
-
-    if (typeof cwd !== "string" || !cwd) {
-      throw new Error("Invalid working directory");
-    }
-    if (!path.isAbsolute(cwd)) {
-      throw new Error("Working directory must be an absolute path");
-    }
-
-    const { listCommits } = await import("../../utils/git.js");
-
-    return listCommits({ cwd, search, branch, skip, limit });
-  };
-  ipcMain.handle(CHANNELS.GIT_LIST_COMMITS, handleGitListCommits);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.GIT_LIST_COMMITS));
-
-  // Task orchestration handlers
-
-  const handleWorktreeCreateForTask = async (
-    event: Electron.IpcMainInvokeEvent,
-    payload: CreateForTaskPayload
-  ): Promise<WorktreeState> => {
-    await waitForRateLimitSlot(WORKTREE_RATE_LIMIT_KEY, WORKTREE_RATE_LIMIT_INTERVAL_MS);
-    if (!deps.worktreeService) {
-      throw new Error("Workspace client not initialized");
-    }
-
-    if (!payload || typeof payload !== "object") {
-      throw new Error("Invalid payload for worktree:create-for-task");
-    }
-
-    const { taskId, baseBranch, description } = payload;
-
-    if (typeof taskId !== "string" || !taskId.trim()) {
-      throw new Error("Invalid taskId: must be a non-empty string");
-    }
-
-    // Get the current project to determine root path
-    const project = projectStore.getCurrentProject();
-
-    if (!project || !project.path) {
-      throw new Error("No active project found");
-    }
-
-    const rootPath = project.path;
-
-    // Get all states to find the main worktree and determine base branch
-    const senderWindowCreate = getWindowForWebContents(event.sender);
-    const states = await deps.worktreeService.getAllStatesAsync(senderWindowCreate?.id);
-    const mainWorktree = states.find((wt) => wt.isMainWorktree);
-
-    // Use provided baseBranch, or default to main worktree's branch, or "main"
-    const effectiveBaseBranch = baseBranch || mainWorktree?.branch || "main";
-
-    // Generate collision-safe branch name: task-{taskId} with suffix if needed
-    const gitService = taskWorktreeService.getGitService(rootPath);
-    const baseBranchName = `task-${taskId}`;
-    const availableBranchName = await gitService.findAvailableBranchName(baseBranchName);
-
-    // Generate path using the worktree path pattern (project-level → global → default)
-    const pattern = await resolveWorktreePattern(rootPath);
-    const initialPath = generateWorktreePath(rootPath, availableBranchName, pattern);
-    const availablePath = gitService.findAvailablePath(initialPath);
-
-    logDebug("Creating worktree for task", {
-      taskId,
-      projectId: project.id,
-      branch: availableBranchName,
-      path: availablePath,
-      baseBranch: effectiveBaseBranch,
-      description,
-    });
-
-    let worktreeId: string;
-    try {
-      // Create the worktree
-      worktreeId = await deps.worktreeService.createWorktree(rootPath, {
-        baseBranch: effectiveBaseBranch,
-        newBranch: availableBranchName,
-        path: availablePath,
-      });
-
-      // Invalidate file search cache for the new worktree path
-      try {
-        fileSearchService.invalidate(availablePath);
-      } catch (error) {
-        console.warn("[worktree.create-for-task] Failed to invalidate file search cache:", error);
-      }
-
-      // Store the taskId mapping
-      taskWorktreeService.addTaskWorktreeMapping(project.id, taskId, worktreeId);
-
-      logDebug("Worktree created for task", {
-        worktreeId,
-        taskId,
-        projectId: project.id,
-        branch: availableBranchName,
-        path: availablePath,
-      });
-    } catch (error) {
-      // Worktree creation failed, don't leave partial state
-      logError(
-        "Failed to create worktree for task",
-        error instanceof Error ? error : new Error(String(error)),
-        {
-          taskId,
-          projectId: project.id,
-        }
-      );
-      throw error;
-    }
-
-    // Wait for the monitor to be ready and return the full state
-    // Retry a few times since the worktree might not be fully indexed yet
-    let state: WorktreeState | null = null;
-    try {
-      for (let i = 0; i < 5; i++) {
-        const monitor = await deps.worktreeService.getMonitorAsync(worktreeId);
-        if (monitor) {
-          state = {
-            id: monitor.id,
-            path: monitor.path,
-            name: monitor.name,
-            branch: monitor.branch,
-            isCurrent: monitor.isCurrent,
-            isMainWorktree: monitor.isMainWorktree,
-            gitDir: monitor.gitDir,
-            summary: monitor.summary,
-            modifiedCount: monitor.modifiedCount,
-            changes: monitor.changes,
-            mood: monitor.mood,
-            lastActivityTimestamp: monitor.lastActivityTimestamp ?? null,
-            createdAt: monitor.createdAt,
-            aiNote: monitor.aiNote,
-            aiNoteTimestamp: monitor.aiNoteTimestamp,
-            issueNumber: monitor.issueNumber,
-            prNumber: monitor.prNumber,
-            prUrl: monitor.prUrl,
-            prState: monitor.prState,
-            prTitle: monitor.prTitle,
-            issueTitle: monitor.issueTitle,
-            worktreeChanges: monitor.worktreeChanges ?? null,
-            worktreeId: monitor.worktreeId,
-            taskId,
-          };
-          break;
-        }
-        // Wait before retrying
-        await new Promise((resolve) => setTimeout(resolve, 200));
-      }
-
-      if (!state) {
-        // Worktree was created but not yet indexed, return minimal state
-        state = {
-          id: worktreeId,
-          path: availablePath,
-          name: availableBranchName,
-          branch: availableBranchName,
-          isCurrent: false,
-          lastActivityTimestamp: null,
-          worktreeId: worktreeId,
-          worktreeChanges: null,
-          taskId,
-        };
-      }
-
-      return state;
-    } catch (error) {
-      // Monitor polling failed - worktree exists but we can't get its state
-      // Return minimal state rather than failing completely
-      logError(
-        "Failed to get monitor state after creating worktree",
-        error instanceof Error ? error : new Error(String(error)),
-        {
-          worktreeId,
-          taskId,
-          projectId: project.id,
-        }
-      );
-
-      return {
-        id: worktreeId,
-        path: availablePath,
-        name: availableBranchName,
-        branch: availableBranchName,
-        isCurrent: false,
-        lastActivityTimestamp: null,
-        worktreeId: worktreeId,
-        worktreeChanges: null,
-        taskId,
-      };
-    }
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_CREATE_FOR_TASK, handleWorktreeCreateForTask);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_CREATE_FOR_TASK));
-
-  const handleWorktreeGetByTaskId = async (
-    _event: Electron.IpcMainInvokeEvent,
-    taskId: string
-  ): Promise<WorktreeState[]> => {
-    if (!deps.worktreeService) {
-      throw new Error("Workspace client not initialized");
-    }
-
-    if (typeof taskId !== "string" || !taskId.trim()) {
-      throw new Error("Invalid taskId: must be a non-empty string");
-    }
-
-    // Get current project to scope the lookup
-    const currentProjectId = projectStore.getCurrentProjectId();
-    if (!currentProjectId) {
-      throw new Error("No active project found");
-    }
-
-    const worktreeIds = taskWorktreeService.getWorktreeIdsForTask(currentProjectId, taskId);
-    const results: WorktreeState[] = [];
-
-    for (const worktreeId of worktreeIds) {
-      const monitor = await deps.worktreeService.getMonitorAsync(worktreeId);
-      if (monitor) {
-        results.push({
-          id: monitor.id,
-          path: monitor.path,
-          name: monitor.name,
-          branch: monitor.branch,
-          isCurrent: monitor.isCurrent,
-          isMainWorktree: monitor.isMainWorktree,
-          gitDir: monitor.gitDir,
-          summary: monitor.summary,
-          modifiedCount: monitor.modifiedCount,
-          changes: monitor.changes,
-          mood: monitor.mood,
-          lastActivityTimestamp: monitor.lastActivityTimestamp ?? null,
-          createdAt: monitor.createdAt,
-          aiNote: monitor.aiNote,
-          aiNoteTimestamp: monitor.aiNoteTimestamp,
-          issueNumber: monitor.issueNumber,
-          prNumber: monitor.prNumber,
-          prUrl: monitor.prUrl,
-          prState: monitor.prState,
-          prTitle: monitor.prTitle,
-          issueTitle: monitor.issueTitle,
-          worktreeChanges: monitor.worktreeChanges ?? null,
-          worktreeId: monitor.worktreeId,
-          taskId,
-        });
-      }
-    }
-
-    return results;
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_GET_BY_TASK_ID, handleWorktreeGetByTaskId);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_GET_BY_TASK_ID));
-
-  const handleWorktreeCleanupTask = async (
-    event: Electron.IpcMainInvokeEvent,
-    taskId: string,
-    options?: CleanupTaskOptions
-  ): Promise<void> => {
-    checkRateLimit(CHANNELS.WORKTREE_CLEANUP_TASK, 10, 10_000);
-    if (!deps.worktreeService) {
-      throw new Error("Workspace client not initialized");
-    }
-
-    if (typeof taskId !== "string" || !taskId.trim()) {
-      throw new Error("Invalid taskId: must be a non-empty string");
-    }
-
-    // Get current project to scope the cleanup
-    const currentProjectId = projectStore.getCurrentProjectId();
-    if (!currentProjectId) {
-      throw new Error("No active project found");
-    }
-
-    const worktreeIds = taskWorktreeService.getWorktreeIdsForTask(currentProjectId, taskId);
-
-    if (worktreeIds.length === 0) {
-      // No worktrees to clean up, return silently (idempotent)
-      return;
-    }
-
-    const force = options?.force ?? true;
-    const deleteBranch = options?.deleteBranch ?? true;
-
-    logDebug("Cleaning up worktrees for task", {
-      taskId,
-      projectId: currentProjectId,
-      worktreeIds,
-      force,
-      deleteBranch,
-    });
-
-    const errors: string[] = [];
-
-    // Fetch all worktree states once for efficient main-worktree checking
-    const senderWindowCleanup = getWindowForWebContents(event.sender);
-    let allStates: Awaited<ReturnType<typeof deps.worktreeService.getAllStatesAsync>> = [];
-    try {
-      allStates = await deps.worktreeService.getAllStatesAsync(senderWindowCleanup?.id);
-    } catch (error) {
-      logDebug("Could not fetch worktree states for cleanup pre-check", {
-        error: error instanceof Error ? error.message : String(error),
-        taskId,
-      });
-    }
-
-    for (const worktreeId of worktreeIds) {
-      try {
-        // Safeguard: Check if this is the main worktree before attempting deletion
-        const targetWorktree = allStates.find((wt) => wt.id === worktreeId);
-
-        if (targetWorktree?.isMainWorktree) {
-          logDebug("Skipping deletion of main worktree in cleanup task", {
-            worktreeId,
-            taskId,
-            projectId: currentProjectId,
-          });
-          taskWorktreeService.removeTaskWorktreeMapping(currentProjectId, taskId, worktreeId);
-          continue;
-        }
-
-        await deps.worktreeService.deleteWorktree(worktreeId, force, deleteBranch);
-        if (targetWorktree) {
-          try {
-            fileSearchService.invalidate(targetWorktree.path);
-          } catch (error) {
-            console.warn("[worktree.cleanup-task] Failed to invalidate file search cache:", error);
-          }
-        }
-
-        // Remove from tracking after successful deletion
-        taskWorktreeService.removeTaskWorktreeMapping(currentProjectId, taskId, worktreeId);
-      } catch (error) {
-        const errorMessage = error instanceof Error ? error.message : String(error);
-
-        // If worktree not found, treat as success and remove mapping (idempotent)
-        if (errorMessage.includes("not found") || errorMessage.includes("does not exist")) {
-          logDebug("Worktree already removed, cleaning up mapping", {
-            worktreeId,
-            taskId,
-            projectId: currentProjectId,
-          });
-          taskWorktreeService.removeTaskWorktreeMapping(currentProjectId, taskId, worktreeId);
-        } else {
-          logError(
-            "Failed to cleanup worktree",
-            error instanceof Error ? error : new Error(errorMessage),
-            {
-              worktreeId,
-              taskId,
-              projectId: currentProjectId,
-            }
-          );
-          errors.push(`${worktreeId}: ${errorMessage}`);
-        }
-      }
-    }
-
-    logDebug("Task cleanup completed", {
-      taskId,
-      projectId: currentProjectId,
-      worktreeIds,
-      errors: errors.length > 0 ? errors : undefined,
-    });
-
-    if (errors.length > 0) {
-      throw new Error(`Failed to cleanup some worktrees: ${errors.join("; ")}`);
-    }
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_CLEANUP_TASK, handleWorktreeCleanupTask);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_CLEANUP_TASK));
-
-  // Issue attachment handlers
-
-  const handleWorktreeAttachIssue = async (
-    _event: Electron.IpcMainInvokeEvent,
-    payload: AttachIssuePayload
-  ): Promise<void> => {
-    if (!payload || typeof payload !== "object") {
-      throw new Error("Invalid payload for worktree:attach-issue");
-    }
-
-    const { worktreeId, issueNumber, issueTitle, issueState, issueUrl } = payload;
-
-    if (typeof worktreeId !== "string" || !worktreeId.trim()) {
-      throw new Error("Invalid worktreeId: must be a non-empty string");
-    }
-    if (typeof issueNumber !== "number" || !Number.isInteger(issueNumber) || issueNumber <= 0) {
-      throw new Error("Invalid issueNumber: must be a positive integer");
-    }
-    if (typeof issueTitle !== "string") {
-      throw new Error("Invalid issueTitle: must be a string");
-    }
-    if (issueState !== "OPEN" && issueState !== "CLOSED") {
-      throw new Error("Invalid issueState: must be 'OPEN' or 'CLOSED'");
-    }
-    if (typeof issueUrl !== "string" || !issueUrl.trim()) {
-      throw new Error("Invalid issueUrl: must be a non-empty string");
-    }
-
-    const association: IssueAssociation = {
-      issueNumber,
-      issueTitle,
-      issueState,
-      issueUrl,
-    };
-
-    const currentMap = store.get("worktreeIssueMap") ?? {};
-    store.set("worktreeIssueMap", { ...currentMap, [worktreeId]: association });
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_ATTACH_ISSUE, handleWorktreeAttachIssue);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_ATTACH_ISSUE));
-
-  const handleWorktreeDetachIssue = async (
-    _event: Electron.IpcMainInvokeEvent,
-    payload: DetachIssuePayload
-  ): Promise<void> => {
-    if (!payload || typeof payload !== "object") {
-      throw new Error("Invalid payload for worktree:detach-issue");
-    }
-
-    const { worktreeId } = payload;
-
-    if (typeof worktreeId !== "string" || !worktreeId.trim()) {
-      throw new Error("Invalid worktreeId: must be a non-empty string");
-    }
-
-    const currentMap = store.get("worktreeIssueMap") ?? {};
-    const { [worktreeId]: _removed, ...rest } = currentMap;
-    store.set("worktreeIssueMap", rest);
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_DETACH_ISSUE, handleWorktreeDetachIssue);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_DETACH_ISSUE));
-
-  const handleWorktreeGetIssueAssociation = async (
-    _event: Electron.IpcMainInvokeEvent,
-    worktreeId: string
-  ): Promise<IssueAssociation | null> => {
-    if (typeof worktreeId !== "string" || !worktreeId.trim()) {
-      throw new Error("Invalid worktreeId: must be a non-empty string");
-    }
-
-    const currentMap = store.get("worktreeIssueMap") ?? {};
-    return currentMap[worktreeId] ?? null;
-  };
-  ipcMain.handle(CHANNELS.WORKTREE_GET_ISSUE_ASSOCIATION, handleWorktreeGetIssueAssociation);
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_GET_ISSUE_ASSOCIATION));
-
-  const handleWorktreeGetAllIssueAssociations = async (): Promise<
-    Record<string, IssueAssociation>
-  > => {
-    return store.get("worktreeIssueMap") ?? {};
-  };
-  ipcMain.handle(
-    CHANNELS.WORKTREE_GET_ALL_ISSUE_ASSOCIATIONS,
-    handleWorktreeGetAllIssueAssociations
-  );
-  handlers.push(() => ipcMain.removeHandler(CHANNELS.WORKTREE_GET_ALL_ISSUE_ASSOCIATIONS));
-
-  return () => {
-    handlers.forEach((cleanup) => cleanup());
-  };
-}
+/**
+ * Worktree handlers - Re-exports from the modular worktree handlers.
+ *
+ * @deprecated Import directly from './worktree/index.js' for new code.
+ * This file is kept for backward compatibility with existing imports.
+ */
+
+export { registerWorktreeHandlers } from "./worktree/index.js";
